refactor publisher and orchestrator
* use dynamodb stream with sns to trigger publisher, removed cloudwatch events.
* use sns subscription filter policy, removed logic from original orchestrator lambda.

+++ orchestrator/pkg/runner/ec2/terminator_test.go
@@ -16,7 +16,7 @@ import (
 
 func TestEc2Terminator_Terminate(t *testing.T) {
 	cases := map[string]struct {
-		id                              int
+		id                              uint64
 		numInstances                    int
 		describeInstancesErr            error
 		expectedDescribeInstanceInput   *ec2.DescribeInstancesInput
